geopublica-2019/requirements.txt
aniso8601==5.1.0
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
entrypoints==0.3
flake8==3.7.7
Flask==1.0.3
Flask-GraphQL==2.0.0
graphene==2.1.5
graphql-core==2.2
graphql-relay==0.4.5
graphql-server-core==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkl-fft==1.0.12
mkl-random==1.0.2
numpy==1.16.4
pandas==0.24.2
promise==2.2.1
psycopg2==2.7.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
relativetimebuilder==1.0.0
requests==2.22.0
Rx==1.6.1
six==1.12.0
toml==0.10.0
urllib3==1.24.2
Werkzeug==0.15.4
